requirements.txt
openai>=1.50.0
python-dotenv>=1.0.0
boto3>=1.35.0
